% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{cross_interactions}
\alias{cross_interactions}
\title{Compute Cross-Interactions}
\usage{
cross_interactions(z1, z2, ident = "")
}
\arguments{
\item{z1}{n x p1 dimensional matrix, each column corresponds to one covariate.}

\item{z2}{n x p2 dimensional matrix, each column corresponds to one covariate.}
}
\value{
n x (p1*p2) Matrix which contains the cross interactions
z1[,i]*z2[,j] for i=1,...,p1 and j=1,...,p2. The column name gives involved
covariates.
}
\description{
\code{cross_interactions} computes all cross interactions, that is, products
of columns, between the matrices z1 nad z2.
}
\section{Authors}{

Alexander Kreiss, LSE, United Kingdom. \email{a.kreiss@lse.ac.uk}

Christoph Rothe, University of Mannheim, Germany
}

\seealso{
\code{\link{fourier_basis}}, \code{\link{interaction_terms}},
\code{\link{HighDim_rd}}
}
