% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{fourier_basis}
\alias{fourier_basis}
\title{Compute Fourier Basis Expansions}
\usage{
fourier_basis(z, order)
}
\arguments{
\item{z}{n x p Matrix, each column corresponds to one covariate.}

\item{order}{The order until which the Fourier expansion shall be computed.}
}
\value{
n x (2 x p x order) Matrix, each column contains an expression of the
form sin(2\emph{pi}r\emph{z\link{,i}) or cos(2}pi\emph{r}z\link{,i}) for r=1,...,order and
j=1,...,p. The column names are of the form "FB j sin r" or "FB j cos r".
}
\description{
\code{fourier_basis} computes Fourier Bases expansions for the covariates in z up to a given order.
}
\section{Authors}{

Alexander Kreiss, LSE, United Kingdom. \email{a.kreiss@lse.ac.uk}

Christoph Rothe, University of Mannheim, Germany
}

\seealso{
\code{\link{HighDim_rd}}, \code{\link{interaction_terms}},
\code{\link{cross_interactions}}
}
