% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{interaction_terms}
\alias{interaction_terms}
\title{Compute Interaction Terms}
\usage{
interaction_terms(z)
}
\arguments{
\item{z}{n x p matrix, each column correpsonds to one covariate}
}
\value{
Matrix with n rows and each columns, corresponds to one interaction,
i.e., z[,i]*z[,j] for i,j=1,...,p and i<j. The column name gives the exact
variables which were interacted.
}
\description{
\code{interaction_terms} computes all interaction terms of the covariates,
that is, the products of columns, provided in z.
}
\section{Authors}{

Alexander Kreiss, LSE, United Kingdom. \email{a.kreiss@lse.ac.uk}

Christoph Rothe, University of Mannheim, Germany
}

\seealso{
\code{\link{fourier_basis}}, \code{\link{HighDim_rd}},
\code{\link{cross_interactions}}
}
